Changed how to move an agent. Added active and passive obstacle avoidance modes
1. The movement of agent is now updated based on an internal waypoint list variable instead of just the agent state.
2. There are now two obstacle avoidance modes.
In active mode, the path finding will be triggered whenever the distance between agent and next tile is small enough(a prediction will be made based on instant speed). It's suitable for always finding the best path in a relative small grid.
In passive mode. The path finding will only be triggerd when previously calculated path is blocked.